pyproject.toml
[tool]
[tool.poetry]
name = "acm_dl_searcher"
version = "0.1.0-alpha.2"
homepage = "https://github.com/ahmed-shariff/acm_dl_searcher"
description = "Top-level package for ACM DL Searcher."
authors = ["Ahmed Shariff <shariff.mfa@outlook.com>"]
readme = "README.rst"
license =  "MIT"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.5',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
]
packages = [
    { include = "acm_dl_searcher" },
    { include = "tests", format = "sdist" },
]

include = ["acm_dl_searcher/templates/search_result.html"]

[tool.poetry.dependencies]
python = "^3.6"

click = "^7.1.2"
minilog = "^1.5"
requests = "^2.24.0"
beautifulsoup4 = "^4.9.3"
bibtexparser = "^1.2.0"
tqdm = "^4.51.0"
tabulate = "^0.8.7"
fuzzysearch = "^0.7.3"
liquidpy = "^0.6.2"


[tool.poetry.dev-dependencies]
bumpversion = "*"
coverage = "*"
flake8 = "*"
invoke = "*"
isort = "*"
pylint = "*"

sphinx = "*"
tox = "*"
yapf = "*"

[tool.poetry.scripts]
acm-dl-searcher = 'acm_dl_searcher.cli:cli'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
